requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
boto3==1.40.40
botocore==1.40.40
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
click==8.3.0
colorama==0.4.6
cryptography==42.0.8
django-allauth==65.11.2
django-environ==0.11.2
django-storages==1.14.6
Django==5.0.14
djangorestframework==3.16.1
idna==3.10
jmespath==1.0.1
kombu==5.5.4
packaging==25.0
pillow==10.4.0
prompt-toolkit==3.0.52
psycopg-binary==3.1.20
psycopg==3.1.20
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Markdown==3.6.1
